% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_blue}
\alias{dhhs_blue}
\title{Hex code for the colour: DHHS Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_blue
}
\description{
#004EA8
}
\keyword{datasets}
