% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_blue1}
\alias{dhhs_blue1}
\title{Hex code for the colour: DHHS Blue1}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_blue1
}
\description{
#E0EAF5
}
\keyword{datasets}
