% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_blue3}
\alias{dhhs_blue3}
\title{Hex code for the colour: DHHS Blue3}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_blue3
}
\description{
#B3CAE5
}
\keyword{datasets}
