% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_blue5}
\alias{dhhs_blue5}
\title{Hex code for the colour: DHHS Blue5}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_blue5
}
\description{
"#4D83C2"
}
\keyword{datasets}
