% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_blue6}
\alias{dhhs_blue6}
\title{Hex code for the colour: DHHS Blue6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_blue6
}
\description{
"#1A60B1"
}
\keyword{datasets}
