% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_green1}
\alias{dhhs_green1}
\title{Hex code for the colour: DHHS Green1}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_green1
}
\description{
#E0EFE9
}
\keyword{datasets}
