% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_green2}
\alias{dhhs_green2}
\title{Hex code for the colour: DHHS Green2}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_green2
}
\description{
#CCE5DB
}
\keyword{datasets}
