% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_green3}
\alias{dhhs_green3}
\title{Hex code for the colour: DHHS Green3}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_green3
}
\description{
#B3D7C9
}
\keyword{datasets}
