% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_green4}
\alias{dhhs_green4}
\title{Hex code for the colour: DHHS Green4}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_green4
}
\description{
#99CAB7
}
\keyword{datasets}
