% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_green6}
\alias{dhhs_green6}
\title{Hex code for the colour: DHHS Green6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_green6
}
\description{
"#1A885D"
}
\keyword{datasets}
