% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_greyscale}
\alias{dhhs_greyscale}
\title{Hex code for the colour: DHHS Greyscale}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_greyscale
}
\description{
#53565A
}
\keyword{datasets}
