% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_greyscale3}
\alias{dhhs_greyscale3}
\title{Hex code for the colour: DHHS Greyscale3}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_greyscale3
}
\description{
#BABBBD
}
\keyword{datasets}
