% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_greyscale5}
\alias{dhhs_greyscale5}
\title{Hex code for the colour: DHHS Greyscale5}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_greyscale5
}
\description{
"#87898C"
}
\keyword{datasets}
