% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_greyscale6}
\alias{dhhs_greyscale6}
\title{Hex code for the colour: DHHS Greyscale6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_greyscale6
}
\description{
"#64676B"
}
\keyword{datasets}
