% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_navy}
\alias{dhhs_navy}
\title{Hex code for the colour: DHHS Navy}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_navy
}
\description{
#201547
}
\keyword{datasets}
