% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_navy2}
\alias{dhhs_navy2}
\title{Hex code for the colour: DHHS Navy2}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_navy2
}
\description{
#C7C5D1
}
\keyword{datasets}
