% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_navy3}
\alias{dhhs_navy3}
\title{Hex code for the colour: DHHS Navy3}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_navy3
}
\description{
#B1ADBF
}
\keyword{datasets}
