% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_navy4}
\alias{dhhs_navy4}
\title{Hex code for the colour: DHHS Navy4}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_navy4
}
\description{
#9B96AC
}
\keyword{datasets}
