% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_navy5}
\alias{dhhs_navy5}
\title{Hex code for the colour: DHHS Navy5}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_navy5
}
\description{
"#635B7E"
}
\keyword{datasets}
