% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_navy6}
\alias{dhhs_navy6}
\title{Hex code for the colour: DHHS Navy6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_navy6
}
\description{
"#362C59"
}
\keyword{datasets}
