% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_orange}
\alias{dhhs_orange}
\title{Hex code for the colour: DHHS Orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_orange
}
\description{
#C5511A
}
\keyword{datasets}
