% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_orange1}
\alias{dhhs_orange1}
\title{Hex code for the colour: DHHS Orange1}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_orange1
}
\description{
#F8EAE4
}
\keyword{datasets}
