% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_orange2}
\alias{dhhs_orange2}
\title{Hex code for the colour: DHHS Orange2}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_orange2
}
\description{
#F3DCD1
}
\keyword{datasets}
