% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_orange4}
\alias{dhhs_orange4}
\title{Hex code for the colour: DHHS Orange4}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_orange4
}
\description{
#E8B9A3
}
\keyword{datasets}
