% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_orange6}
\alias{dhhs_orange6}
\title{Hex code for the colour: DHHS Orange6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_orange6
}
\description{
"#CB6231"
}
\keyword{datasets}
