% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhhs_pal.R
\name{dhhs_pal}
\alias{dhhs_pal}
\title{Create a DHHS-appropriate palette for your chart.}
\usage{
dhhs_pal(n = 0, reverse = FALSE, faded = FALSE)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is
  1, maximum is 10. Using more than 6 is not recommended. If you don't
  specify `n`, a five-colour palette will be used, which may not look right.
  Specify `n`.

  By default, n = 2 will give you teal and red.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order. Standard colour order runs from light to dark. If
you set reverse to TRUE, colours will run from dark to light.}

\item{faded}{Logical. FALSE by default. Setting to TRUE returns the faded
variations of the standard colours.}
}
\description{
Create a DHHS-appropriate palette for your chart.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_dhhs() +
    scale_colour_manual(values = dhhs_pal(n = 3))

p

# Alternatively, use dhhs_colour_manual(), which is a wrapper
# around scale_colour_manual():

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_dhhs() +
    dhhs_colour_manual(n = 3)

p

}
