% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_palette_set}
\alias{dhhs_palette_set}
\title{A list of DHHS colour sets}
\format{
An object of class \code{list} of length 10.
}
\usage{
dhhs_palette_set
}
\description{
"teal": from light teal to dark teal
"blue": from light blue to dark blue
"pink": from light pink to dark pink
"green": from light green to dark green
"navy": from light navy to dark navy
"purple": from light purple to dark purple
"orange": from light orange to dark orange
"diverging": pink, faded pink, white, faded teal, teal
"diverging2": pink, faded pink, white, faded navy, navy
"grey": from light greyscale to dark greyscale
}
\keyword{datasets}
