% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_pink}
\alias{dhhs_pink}
\title{Hex code for the colour: DHHS Pink}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_pink
}
\description{
#D50032
}
\keyword{datasets}
