% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_pink1}
\alias{dhhs_pink1}
\title{Hex code for the colour: DHHS Pink1}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_pink1
}
\description{
#FAE0E6
}
\keyword{datasets}
