% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_pink2}
\alias{dhhs_pink2}
\title{Hex code for the colour: DHHS Pink2}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_pink2
}
\description{
#F7CCD6
}
\keyword{datasets}
