% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_pink4}
\alias{dhhs_pink4}
\title{Hex code for the colour: DHHS Pink4}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_pink4
}
\description{
#EE99AD
}
\keyword{datasets}
