% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_pink5}
\alias{dhhs_pink5}
\title{Hex code for the colour: DHHS Pink5}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_pink5
}
\description{
"#E24D70"
}
\keyword{datasets}
