% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_pink6}
\alias{dhhs_pink6}
\title{Hex code for the colour: DHHS Pink6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_pink6
}
\description{
"#D91A47"
}
\keyword{datasets}
