% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple}
\alias{dhhs_purple}
\title{Hex code for the colour: DHHS Purple}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple
}
\description{
#87189D
}
\keyword{datasets}
