% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple1}
\alias{dhhs_purple1}
\title{Hex code for the colour: DHHS Purple1}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple1
}
\description{
#F1E3F3
}
\keyword{datasets}
