% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple2}
\alias{dhhs_purple2}
\title{Hex code for the colour: DHHS Purple2}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple2
}
\description{
#E7D1EB
}
\keyword{datasets}
