% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple3}
\alias{dhhs_purple3}
\title{Hex code for the colour: DHHS Purple3}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple3
}
\description{
#DBBAE2
}
\keyword{datasets}
