% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple4}
\alias{dhhs_purple4}
\title{Hex code for the colour: DHHS Purple4}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple4
}
\description{
#CFA3D8
}
\keyword{datasets}
