% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple5}
\alias{dhhs_purple5}
\title{Hex code for the colour: DHHS Purple5}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple5
}
\description{
"#AB5DBA"
}
\keyword{datasets}
