% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_purple6}
\alias{dhhs_purple6}
\title{Hex code for the colour: DHHS Purple6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_purple6
}
\description{
"#932FA7"
}
\keyword{datasets}
