% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhhs_save.R
\name{dhhs_save}
\alias{dhhs_save}
\title{A function to export ggplot2 objects in consistent DHHS style}
\usage{
dhhs_save(
  file_path,
  plot_object = ggplot2::last_plot(),
  type = "whole",
  export_chartdata = FALSE,
  drop_labs = FALSE,
  mute_messages = FALSE,
  ...
)
}
\arguments{
\item{file_path}{Save to path, including a type suffix (e.g. ".png" or ".pdf").}

\item{plot_object}{ggplot2 object to plot. Defaults to}

\item{type}{Sets height and width to DHHS defaults. If more than one is
provided, a folder containing the plots will be created. "all" will export all.
The following chart types are available:
\itemize{
\item{"whole"}{ The default. Use for a plot covering the whole body of a
normal DHHS slide. Width: 32.74cm, height: 18.9cm.}
\item{"half"}{ Use for a tall plot covering the full left or right side a normal DHHS slide.
Width: 15.65cm, height: 18.9cm.}
\item{"third"}{ Use for a tall plot covering roughly one-third of the horizontal space on a normal DHHS slide.
Width: 10cm, height: 18.9cm.}
\item{"short-whole"}{ The default. Use for a short plot covering half the body of a
normal DHHS slide. Width: 32.74cm, height: 9.5cm.}
\item{"short-half"}{Use for a short plot covering half of the left or right side of a normal DHHS slide.
Width: 15.65cm, height: 9.5cm.}
\item{"short-third"}{ Use for a short plot covering roughly one-third of the horizontal space on a normal DHHS slide.
Width: 10cm, height: 9.5cm.}
}
Set type = "all" to save your chart in all available sizes.}

\item{export_chartdata}{Export chart data in an Excel file along with the images. Will be set to TRUE if type = "all".}

\item{drop_labs}{Drops title, subtitle and caption. FALSE by default. Useful if you want to write your own in your export destination.}

\item{mute_messages}{Do not display saving messages.}

\item{...}{Other arguments passed to ggsave}
}
\description{
A function to export ggplot2 objects in consistent DHHS style
}
