% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhhs_scales.R
\name{dhhs_scale}
\alias{dhhs_colour_manual}
\alias{dhhs_fill_manual}
\title{Convenient functions to set DHHS-appropriate palettes}
\usage{
dhhs_colour_manual(
  n = NULL,
  reverse = FALSE,
  discrete = TRUE,
  faded = FALSE,
  palette = "teal",
  ...
)

dhhs_fill_manual(
  n = NULL,
  reverse = FALSE,
  discrete = TRUE,
  faded = FALSE,
  palette = "teal",
  ...
)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is
1, maximum is 7. Passed to \code{dhhs_pal}; see ?dhhs_pal for more
information. Set to NULL if you want to provide your own colours with \code{values = c()}.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order.}

\item{discrete}{Logical. TRUE by default. Setting to FALSE generates a
continuous colour scale.}

\item{faded}{Logical. FALSE by default. Setting to TRUE returns faded
variations of the standard colours.}

\item{palette}{Sets the colours that will form the continuous palette when
  discrete = FALSE. One of:

#' A list of DHHS colour sets


\itemize{
\item{"teal"}{from light teal to dark teal}
\item{"blue"}{from light blue to dark blue}
\item{"pink"}{from light pink to dark pink}
\item{"green"}{from light green to dark green}
\item{"navy"}{from light navy to dark navy}
\item{"purple"}{from light purple to dark purple}
\item{"orange"}{from light orange to dark orange}
\item{"diverging"}{pink, faded pink, white, faded teal, teal}
\item{"diverging2"}{pink, faded pink, white, faded navy, navy}
\item{"grey"}{from light greyscale to dark greyscale}
}}

\item{...}{arguments passed to ggplot2 scales}
}
\description{
Convenient functions to set DHHS-appropriate palettes
}
\examples{

library(ggplot2)

ggplot(data = mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
   geom_point() +
   dhhs_colour_manual(n = 3) +
   theme_dhhs()

}
