% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal}
\alias{dhhs_teal}
\title{Hex code for the colour: DHHS Teal}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_teal
}
\description{
#00A3B6
}
\keyword{datasets}
