% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal1}
\alias{dhhs_teal1}
\title{Hex code for the colour: DHHS Teal 1}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_teal1
}
\description{
#00A3B6
}
\keyword{datasets}
