% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal2}
\alias{dhhs_teal2}
\title{Hex code for the colour: DHHS Teal 2}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
dhhs_teal2

dhhs_teal2
}
\description{
#66C8D3

#99DAE1
}
\keyword{datasets}
