% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal3}
\alias{dhhs_teal3}
\title{Hex code for the colour: DHHS Teal 3}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_teal3
}
\description{
#CCECF0
}
\keyword{datasets}
