% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal4}
\alias{dhhs_teal4}
\title{Hex code for the colour: DHHS Teal 4}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_teal4
}
\description{
#E5F5F7
}
\keyword{datasets}
