% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal5}
\alias{dhhs_teal5}
\title{Hex code for the colour: DHHS Teal5}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_teal5
}
\description{
"#4DBFCC"
}
\keyword{datasets}
