% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dhhs_teal6}
\alias{dhhs_teal6}
\title{Hex code for the colour: DHHS Teal6}
\format{
An object of class \code{character} of length 1.
}
\usage{
dhhs_teal6
}
\description{
"#1AACBD"
}
\keyword{datasets}
