% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{owid_sample}
\alias{owid_sample}
\title{A COVID-19 data subset from Our World In Data: https://ourworldindata.org/coronavirus}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1719 rows and 12 columns.
}
\usage{
owid_sample
}
\description{
This full dataset can be downloaded from: https://github.com/owid/covid-19-data/tree/master/public/data
}
\keyword{datasets}
