% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_chartdata.R
\name{save_chartdata}
\alias{save_chartdata}
\title{Save a ggplot2 chart as a properly-formatted "chart data" Excel spreadsheet}
\usage{
save_chartdata(
  filename,
  object = ggplot2::last_plot(),
  type = "whole",
  export_vars = "used",
  add_vars = NULL
)
}
\arguments{
\item{filename}{filename of the Excel workbook you wish to create. Can be a
file path, but the directory must already exist.}

\item{object}{ggplot2 chart to create chart data from. If left blank,
\code{ggplot2::last_plot()} is used to get the last plot displayed.}

\item{type}{type of plot. Default is "normal". See \code{?dhhs_save} for
full list of types.}

\item{export_vars}{Which variables to export data for. Defaults to "used",
which will only export variables used in the plot. Input "all" to include all
variables in the data. Otherwise provide a character vector of variable names.}

\item{add_vars}{A character vector of variables to include \textit{as well as}
the variables in \code{export_vars}.}
}
\description{
This function takes a ggplot2 object and creates a single-sheet Excel workbook
with the chart data and chart. If your ggplot2 object has a subtitle and
caption, those will be properly displayed in the spreadsheet.
}
\examples{

library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
     geom_point() +
     theme_dhhs() +
     labs(title = "Title",
          subtitle = "Subtitle",
          caption = "Caption")

\dontrun{save_chartdata("my_chartdata.xlsx", p)}

}
