% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dhhs.R
\name{theme_dhhs}
\alias{theme_dhhs}
\title{A ggplot2 theme consistent with DHHS style.}
\usage{
theme_dhhs(
  base_size = 14,
  background = "white",
  legend = "none",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 14.}

\item{background}{"white" by default. Set to "transparent" or a different
DHHS colour.}

\item{legend}{"off" by default. Set to "bottom", "left", "right" or "top" as
desired, or a two element numeric vector such as c(0.9, 0.1).}

\item{panel_borders}{`FALSE` by default. Set to `TRUE` to enable a black
border around the plotting area.}
}
\description{
A ggplot2 theme consistent with DHHS style.
}
