% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dhhs.R
\name{theme_dhhs}
\alias{theme_dhhs}
\title{A ggplot2 theme consistent with DHHS style.}
\usage{
theme_dhhs(
  base_size = 14,
  base_colour = dhhs_navy,
  background = "white",
  legend = "right",
  panel_borders = "axes",
  include_minor_gridlines = "none"
)
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 14.}

\item{base_colour}{Colour scheme for titles and axes. One of "navy" (default), "pink", "blue", "green", "orange".}

\item{background}{"white" by default. Set to a different DHHS colour.}

\item{legend}{"right" by default. Set to "none", "bottom", "left", "right" or "top" as
desired, or a two element numeric vector such as c(0.9, 0.1).}

\item{panel_borders}{One of `"axes"` (default), "border", "none". Sets the border to a darker colour.}

\item{include_minor_gridlines}{If "none" (the default), minor gridlines are not included.
Use "x" or "y" to include minor gridlines for x or y axes, or "both" for both axes.}
}
\description{
A ggplot2 theme consistent with DHHS style.
}
